[main] refs #234 - Altcoin manager signing methods in separate file
